single Plot component
